后台酒庄接口开发

+++ src/main/java/com/myicellar/digitalmenu/controller/manage/WineryManageController.java
@@ -2,12 +2,14 @@ package com.myicellar.digitalmenu.controller.manage;
 
 import com.aliyuncs.utils.StringUtils;
 import com.myicellar.digitalmenu.dao.entity.Winery;
+import com.myicellar.digitalmenu.service.WineService;
 import com.myicellar.digitalmenu.service.WineryService;
 import com.myicellar.digitalmenu.shiro.AuthIgnore;
 import com.myicellar.digitalmenu.utils.BizException;
 import com.myicellar.digitalmenu.utils.ConvertUtils;
 import com.myicellar.digitalmenu.utils.SnowflakeIdWorker;
 import com.myicellar.digitalmenu.vo.request.WineryDeleteReqVO;
+import com.myicellar.digitalmenu.vo.request.WineryNameReqVO;
 import com.myicellar.digitalmenu.vo.request.WineryPageReqVO;
 import com.myicellar.digitalmenu.vo.request.WineryReqVO;
 import com.myicellar.digitalmenu.vo.response.PageResponseVO;
@@ -24,6 +26,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Date;
+import java.util.List;
 
 @RestController
 @Slf4j
@@ -33,6 +36,8 @@ public class WineryManageController {
 
     @Autowired
     private WineryService wineryService;
+    @Autowired
+    private WineService wineService;
     @Autowired
     private SnowflakeIdWorker snowflakeIdWorker;
 
@@ -48,6 +53,25 @@ public class WineryManageController {
     public ResultVO<PageResponseVO<WineryRespVO>> queryListPage(@RequestBody WineryPageReqVO reqVO) {
         PageResponseVO<Winery> page = wineryService.queryPageList(reqVO);
 
+        PageResponseVO<WineryRespVO> resultPage = new PageResponseVO<WineryRespVO>();
+        if(page!=null && !CollectionUtils.isEmpty(page.getItems())){
+            resultPage = ConvertUtils.convertPage(page,WineryRespVO.class);
+        }
+
+        return ResultVO.success(resultPage);
+    }
+
+    /**
+     * 详情查询-按酒庄名称
+     *
+     * @param reqVO
+     * @return
+     */
+    @PostMapping(value = "/queryByName")
+    @AuthIgnore
+    @ApiOperation("详情查询-按酒庄名称")
+    public ResultVO<PageResponseVO<WineryRespVO>> queryByName(@RequestBody WineryNameReqVO reqVO) {
+        PageResponseVO<Winery> page = wineryService.queryByName(reqVO);
         PageResponseVO<WineryRespVO> resultPage = new PageResponseVO<WineryRespVO>();
         if(page!=null && !CollectionUtils.isEmpty(page.getItems())){
             resultPage = ConvertUtils.convertPage(page,WineryRespVO.class);
@@ -124,6 +148,10 @@ public class WineryManageController {
         if(reqVO.getWineryId()==null || reqVO.getWineryId()==0L){
             return ResultVO.validError("parameter is invalid！");
         }
+        //如果酒庄里有关联酒品, 无法删除酒庄
+        if (wineService.queryByWineryId(reqVO.getWineryId())!=null){
+            return ResultVO.validError("delete is failed! winery has wine in it.");
+        }
         int i = wineryService.deleteByPrimaryKey(reqVO.getWineryId());
         if(i==0){
             return ResultVO.validError("delete is failed!");
@@ -140,6 +168,8 @@ public class WineryManageController {
         if(reqVO.getWineryId()==null || reqVO.getWineryId()==0L){
             throw new BizException("winery cannot be empty!");
         }
+        //判断酒庄名字是否已经存在
+        this.checkWineyName(reqVO.getWineryNameEng());
         if(StringUtils.isEmpty(reqVO.getWineryNameEng())){
             throw new BizException("wineryNameEng cannot be empty!");
         }
@@ -153,9 +183,26 @@ public class WineryManageController {
         if(reqVO.getWineryId()==null || reqVO.getWineryId()==0L){
             throw new BizException("wineryId cannot be empty!");
         }
+        //判断酒庄名字是否已经存在
+        this.checkWineyName(reqVO.getWineryNameEng());
         if(StringUtils.isEmpty(reqVO.getWineryNameEng())){
             throw new BizException("wineryNameEng cannot be empty!");
         }
     }
+
+    /**
+     * 判断酒庄名字是否存在
+     * @param wineryName
+     */
+    private void checkWineyName(String wineryName){
+        //判断酒庄名字是否已经存在
+        WineryNameReqVO wineryNameReqVO= new WineryNameReqVO();
+        wineryNameReqVO.setWineryNameEng(wineryName);
+        if (wineryService.queryByName(wineryNameReqVO)!=null){
+            throw new BizException("winery already exists");
+        }
+    }
+
+
 
 }
